refactor : Change folder structure and names
