+++ lib/components/widget/sign_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'custom/custom_next_page_button.dart';
+import '../custom_ui_kit/custom_next_page_button.dart';
 
 class SignWidget extends StatelessWidget {
   final String title;
