+++ lib/page/home/home_page.dart
@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import 'package:lister/page/search/search_page.dart';
-import '../../widget/custom/custom_search_bar.dart';
-import '../../widget/list/home_list_view.dart';
+import '../../components/custom_ui_kit/custom_search_bar.dart';
+import 'main_list/home_list_view.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -117,7 +116,7 @@ class HomePage extends StatelessWidget {
 
                 SizedBox(
                   height: 282,
-                  child: HomeListView(),
+                  child: MainListView(),
                 )
               ],
             ),
