+++ lib/page/join/email_authentication_page.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../widget/custom/custom_next_page_button.dart';
-import '../../widget/custom/custom_progress_bar.dart';
-import '../../widget/custom/custom_text_form_field.dart';
-import '../../widget/sign_widget.dart';
+import '../../components/custom_ui_kit/custom_next_page_button.dart';
+import '../../components/custom_ui_kit/custom_progress_bar.dart';
+import '../../components/custom_ui_kit/custom_text_form_field.dart';
+import '../../components/widget/sign_widget.dart';
 import '../../services/api_service.dart';
 import 'set_id_name_page.dart';
