+++ lib/page/join/set_password_page.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 
+import '../../components/custom_ui_kit/custom_next_page_button.dart';
+import '../../components/custom_ui_kit/custom_progress_bar.dart';
+import '../../components/custom_ui_kit/custom_text_form_field.dart';
+import '../../components/pop_up/agreement_pop_up.dart';
+import '../../components/widget/sign_widget.dart';
 import '../../services/api_service.dart';
-import '../../widget/custom/custom_text_form_field.dart';
-import '../../widget/sign_widget.dart';
-import '../../widget/custom/custom_next_page_button.dart';
-import '../../widget/custom/custom_progress_bar.dart';
-import '../../widget/pop_up/agreement_pop_up.dart';
+
 import 'sign_up_congratulation_page.dart';
 
 class SetPasswordPage extends StatefulWidget {
