+++ lib/page/join/sign_up_congratulation_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../../widget/custom/custom_next_page_button.dart';
+import '../../components/custom_ui_kit/custom_next_page_button.dart';
 import '../login/login_page.dart';
 
 class SignUpCongratulationPage extends StatefulWidget {
