+++ lib/page/splash/splash_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:jwt_decoder/jwt_decoder.dart';
 
-import '../../widget/navigator/home_page_navigator.dart';
+import '../../components/navigator/home_page_navigator.dart';
 import '../login/login_page.dart';
 
 class SplashPage extends StatefulWidget {
